requirements.txt
mcp
fastmcp
pydantic
python-dotenv
pymongo
uvicorn
fastapi
mcp.client
langchain-mcp-adapters